// admin/src/apis/proto/openauth/v1/permission_assignments.ts
export interface EffectivePermission {
  /** Permission details */
  permissionId: string;
  permissionName: string;
  permissionDisplayName: string;
  permissionDescription?:
    | string
    | undefined;
  /** Source of the permission: "direct" or "group" */
  source: string;
  /** If source is "group", this contains the group details */
  groupId?: string | undefined;
  groupName?: string | undefined;
  groupDisplayName?:
    | string
    | undefined;
  /** If source is "direct", this may contain expiration info */
  expiresAt?:
    | string
    | undefined;
  /** When this permission was granted */
  grantedAt: string;
  /** Who granted this permission */
  grantedBy: string;
}